requirements.txt
Flask==2.3.3
scikit-learn==1.3.2
nltk==3.8.1
numpy==1.24.3
pandas==2.0.3
textstat==0.7.3
wordcloud==1.9.2
matplotlib==3.7.2
seaborn==0.12.2
plotly==5.17.0
openai==1.107.0
supabase==2.18.1
python-dotenv==1.1.1
tqdm==4.67.1
langdetect==1.0.9
backoff==2.2.1
requests==2.32.5
